created util function for reuse
