+++ src/ExperienceSection.tsx
@@ -1,5 +1,6 @@
 import React, { useState } from 'react';
 import Company from './Company';
+import parseKeywords from './utils/ParseKeywords'
 import './styles/ExperienceSection.scss';
 
 const ExperienceSection = () => {
@@ -8,21 +9,23 @@ const ExperienceSection = () => {
             "company": "Rivian",
             "position": "Software Engineer Intern",
             "time": "Sep 2021 - Apr 2022",
-            "experience": ["Develop and deploy 4 responsive web applications from ground-up in React, Node.js, HTML/CSS, MongoDB, and AWS (S3, EC2), which reduced robot commissioning time by ~75% and increased production in vehicles", "Design user interface in Figma for 2 web applications"]
+            "experience": ["Develop and deploy 4 responsive web applications from ground-up in React, Node.js, HTML, CSS/Sass, MongoDB, and AWS (S3, EC2), which reduced robot commissioning time by ~75% and increased production in vehicles", "Design user interface in Figma for 2 web applications"],
+            "keywords": ["React", "Node.js", "HTML", "CSS", "Sass", "MongoDB", "AWS (S3, EC2)", "Figma"]
         },
         {
             "company": "Oracle Cloud Infrastructure",
             "position": "Software Engineer Intern",
             "time": "May 2021 - Aug 2021",
             "experience": ["Implemented new features, fixed bugs, and created unit tests for the Data Integration Service web page using React, Jest, and the DOM Testing Library",
-            "Built a metrics dashboard with 8 visualizations that monitors site performance and customer usage of the Data Integration Service on the Oracle Cloud Infrastructure platform using Grafana to assist 80+ on-call engineers resolve customer issues"]
+            "Built a metrics dashboard with 8 visualizations that monitors site performance and customer usage of the Data Integration Service on the Oracle Cloud Infrastructure platform using Grafana to assist 80+ on-call engineers resolve customer issues"],
+            "keywords": ["React", "Jest", "DOM Testing Library", "Grafana"]
         },
         {
             "company": "USC",
             "position": "Software Engineering (CS 310) Course Producer",
             "time": "Aug 2020 - Dec 2020",
             "experience": ["Assist with the understanding of software engineering concepts, including project management, architecture, design, implementation, testing, and maintenance for 150+ students", 
-            "Serve as stakeholder and mentor in technical software engineering group projects"]
+            "Serve as stakeholder and mentor in technical software engineering group projects"],
         },
         {
             "company": "Dell Technologies",
@@ -30,7 +33,8 @@ const ExperienceSection = () => {
             "time": "Jun 2020 - Aug 2020",
             "experience": ["Automate unit and integration testing for data operations team by writing scripts to integrate Github and Jenkins and configuring slave nodes",
             "Create a Jenkins Continuous Integration/Continuous Deployment pipeline by bundling data operations and tests into a Docker container and deploying it to Kubernetes",
-            "Write and fix python unit tests for ETL operations"]
+            "Write and fix Python unit tests for ETL operations"],
+            "keywords": ["Github", "Jenkins", "Docker", "Kubernetes", "Python"]
         },
         {
             "company": "USC",
@@ -44,7 +48,8 @@ const ExperienceSection = () => {
             "position": "Instructor",
             "time": "Jun 2019 - Aug 2019",
             "experience": ["Created curriculum and engaging activities to teach up to 10 students per week between the ages of 10 to 18 years old programming in C++ and Java and game development on Roblox",
-            "Assisted students in creating personal projects from scratch using the C++ and Java programming language"]
+            "Assisted students in creating personal projects from scratch using the C++ and Java programming language"],
+            "keywords": ["C++", "Java"]
         }
     ]
 
@@ -75,7 +80,8 @@ const ExperienceSection = () => {
             <div className="time"> {experienceArray[clickedExp].time}</div>
             <ul className="details"> 
             {experienceArray[clickedExp].experience.map((bullet) => {
-                return <li className="bullet">{ bullet }</li>
+                const str = parseKeywords(experienceArray[clickedExp].keywords, bullet);
+                return <li className="bullet" dangerouslySetInnerHTML={{__html: str}}></li>
             })
             }
             </ul>
