+++ src/styles/ExperienceSection.scss
@@ -37,6 +37,11 @@
                 padding-left: 20px;
                 .bullet {
                     margin-bottom: 10px;
+                    span {
+                        display: inline;
+                        color: #FECB02;
+                        font-weight: 700;
+                    }
                 }
                
             }
